toasty/pyramid.py: start breaking non-TOAST-specific tile stuff into this module
