+++ docs/api.rst
@@ -17,6 +17,10 @@ API Documentation
    :no-inheritance-diagram:
    :no-inherited-members:
 
+.. automodapi:: toasty.pyramid
+   :no-inheritance-diagram:
+   :no-inherited-members:
+
 .. automodapi:: toasty.samplers
    :no-inheritance-diagram:
    :no-inherited-members:
