+++ docs/api/toasty.pyramid.pos_parent.rst
@@ -1,0 +1,6 @@
+pos_parent
+==========
+
+.. currentmodule:: toasty.pyramid
+
+.. autofunction:: pos_parent
